Add tests

+++ app/models/users.go
@@ -8,7 +8,7 @@ const (
 )
 
 type User struct {
-	ID      int64  `db:"id"`
+	ID      int    `db:"id"`
 	Email   string `db:"email"`
 	Pwdhash string `db:"pwdhash"`
 	Role    Role   `db:"role"`
